Rename email-authenticator files.
