+++ packages/authentication/src/email-and-password/email-and-password-authenticator.service.ts
