+++ packages/authentication/src/email-and-password/index.ts
@@ -1,2 +1,2 @@
-export { CheckPassword, EmailAuthenticatorService } from './email-authenticator.service';
+export { CheckPassword, EmailAuthenticatorService } from './email-and-password.authenticator.service';
 export { validateEmailCredentialsFormat } from './validate-email-credentials-format.pre-hook';
